Add ability to create, bind and use RWDataBuffers
All the metadata generation is also complete
Rework GPU Buffer creation to account for data format

+++ Source/Rendering/CrRendering.h
@@ -8,7 +8,7 @@ namespace cr3d
 
 	namespace DataFormat
 	{
-		enum T : uint32_t
+		enum T : uint8_t
 		{
 			//-------------
 			// Uncompressed
